+++ PCN/PCN.h
@@ -119,7 +119,7 @@ public:
     void SetTrackingPeriod(int period);
     int GetTrackingPeriod();
     void SetTrackingThresh(float thresh);
-    std::vector<Window> GenerateThirdLayerInput(cv::Mat img, db::DB *db, int image_label);
+    std::vector<Window> GenerateThirdLayerInput(cv::Mat img, db::DB *db, float image_label);
     float ProcessSingleImageStage3(std::vector<cv::Mat> image);
     /// detection
     std::vector<Window> Detect(cv::Mat img);
@@ -161,8 +161,8 @@ private:
     std::vector<Window> Detect_(cv::Mat img, cv::Mat imgPad);
     std::vector<Window> Track_(cv::Mat img, caffe::shared_ptr<caffe::Net<float> > &net,
                                float thres, int dim, std::vector<Window> &winList);
-    void DumpImage_(std::vector<cv::Mat> &input, db::DB *db, int image_label);
-    void PartialStage3_(cv::Mat img, cv::Mat img180, cv::Mat img90, cv::Mat imgNeg90, caffe::shared_ptr<caffe::Net<float> > &net, float thres, int dim, std::vector<Window> &winList, db::DB *db, int image_label);
+    void DumpImage_(std::vector<cv::Mat> &input, db::DB *db, float image_label);
+    void PartialStage3_(cv::Mat img, cv::Mat img180, cv::Mat img90, cv::Mat imgNeg90, caffe::shared_ptr<caffe::Net<float> > &net, float thres, int dim, std::vector<Window> &winList, db::DB *db, float image_label);
     float OnlyStage3_(std::vector<cv::Mat> dataList, caffe::shared_ptr<caffe::Net<float> > &net);
 
     //private data
